requirements.txt
python-memcached==1.58
redis==2.10.5



